+++ src/types/theme.d.ts
@@ -82,7 +82,6 @@ declare module '@emotion/react' {
         info: { 500: string };
         warning: { 500: string };
         danger: { 500: string };
-        lightblue: { 500: string };
       };
     };
     // spacing: {
